+++ packages/web-components/src/components.d.ts
@@ -19768,7 +19768,11 @@ export namespace Components {
         "end": string;
         "interval": any;
         /**
-          * Display the day (MM/DD) at 00:00. Only works when Timeline interval is set to 'hour'.
+          * Display a secondary ruler which shows one level up from the current interval. IE: Days for Hours, Months for Days, etc.
+         */
+        "showSecondaryRuler"?: boolean | undefined;
+        /**
+          * Display the day (MM/DD) at 00:00. Only works when Timeline interval is set to 'hour' or 'minutes'.
          */
         "showStartOfDay"?: boolean | undefined;
         "start": string;
@@ -20086,6 +20090,10 @@ export namespace Components {
           * The timeline's playhead date time. Must be an ISO string "2021-02-02T05:00:00Z"
          */
         "playhead"?: string;
+        /**
+          * Controls the display of grid lines
+         */
+        "showGrid": boolean;
         /**
           * The timeline's start date. Must be an ISO string "2021-02-02T05:00:00Z"
          */
@@ -55342,7 +55350,11 @@ declare namespace LocalJSX {
         "end"?: string;
         "interval"?: any;
         /**
-          * Display the day (MM/DD) at 00:00. Only works when Timeline interval is set to 'hour'.
+          * Display a secondary ruler which shows one level up from the current interval. IE: Days for Hours, Months for Days, etc.
+         */
+        "showSecondaryRuler"?: boolean | undefined;
+        /**
+          * Display the day (MM/DD) at 00:00. Only works when Timeline interval is set to 'hour' or 'minutes'.
          */
         "showStartOfDay"?: boolean | undefined;
         "start"?: string;
@@ -55713,6 +55725,10 @@ declare namespace LocalJSX {
           * The timeline's playhead date time. Must be an ISO string "2021-02-02T05:00:00Z"
          */
         "playhead"?: string;
+        /**
+          * Controls the display of grid lines
+         */
+        "showGrid"?: boolean;
         /**
           * The timeline's start date. Must be an ISO string "2021-02-02T05:00:00Z"
          */