+++ packages/framework-testing/astro-react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # astro-react
 
+## 0.0.0-timeline-enhanced-3
+
+### Patch Changes
+
+- fixed an issue with secondary ruler and date not working if start time minute was anything but 00:00
+
+## 0.0.0-timeline-enhanced-20240531164326
+
+### Minor Changes
+
+- ac4380a7: feat(rux-timeline) added a showSecondaryRuler prop to rux-ruler to show a secondary rule that is one increment up from the main interval (for days it shows months, for months it shows years, etc.)
+- 023b1de1: feat(rux-timeline) add a showGrid option to rux-timeline when implemented it displays gridlines on the timeline
+
 ## 0.3.1
 
 ### Patch Changes